♻️(front) move DashboardButtons to a dedicated module
DashboardButton and derived components are moved to a dedicated module.
We made this to avoid loading lot of unwanted libs when we want to use
these buttons.
